Update button component file structure
